Move dog models to dogs catalog
