+++ src/app/modules/dogs/models/dog.model.ts
