+++ src/app/modules/dogs/models/training.model.ts
